rename pattern-4 to patter4
