+++ pkg/controller/mysql/deployment.go
@@ -21,10 +21,10 @@ package mysql
 
 import (
 	apimv1alpha1 "github.com/wso2/k8s-wso2am-operator/pkg/apis/apim/v1alpha1"
-	pattern4 "github.com/wso2/k8s-wso2am-operator/pkg/controller/pattern-4"
 	"github.com/wso2/k8s-wso2am-operator/pkg/controller/pattern1"
 	"github.com/wso2/k8s-wso2am-operator/pkg/controller/pattern2"
 	"github.com/wso2/k8s-wso2am-operator/pkg/controller/pattern3"
+	"github.com/wso2/k8s-wso2am-operator/pkg/controller/pattern4"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
