+++ pkg/controller/pattern4/config.go
