+++ pkg/controller/pattern4/constants.go
