+++ pkg/controller/pattern4/deployment.go
