+++ pkg/controller/pattern4/initcontainers.go
