+++ pkg/controller/pattern4/service.go
